Bump version to 1.0.0
- Fixed dedicated server feeding: mod now runs on the ZDO owner (client)
  rather than assuming server-only execution
- Replaced Physics.OverlapSphere container discovery with Piece.s_allPieces
  iteration, fixing containers not being found on clients due to collider
  batching
- Replaced custom server RPC with direct container.Save() inventory sync,
  matching the original 2024 approach that correctly propagates via ZNetView
- Removed dead ChestLayer constant from PluginSettings
- Removed ColliderExtensions (superseded by Vector3Extensions)
- Added CHANGELOG.md to Thunderstore package
