--- Extensions/ColliderExtensions.cs
@@ -1,33 +1,0 @@
-namespace AutoFeed;
-
-public static class ColliderExtensions
-{
-    public static List<Container> GetContainersOrderedByDistanceFromCenter(
-        this Collider[] colliders,
-        Vector3 center
-    )
-    {
-        var containers = colliders
-            .Select(collider => collider.GetComponentInParent<Container>())
-            .Where(container =>
-                container is not null
-                && IsValidZNetView(container.GetComponent<ZNetView>())
-                && IsNonEmptyChest(container)
-            )
-            .OrderBy(container => Vector3.Distance(container.transform.position, center));
-        return [.. containers];
-    }
-
-    private static bool IsValidZNetView(ZNetView? zNetView) =>
-        zNetView is not null && zNetView.IsValid();
-
-    private static bool IsNonEmptyChest(Container container)
-    {
-        var inventory = container.GetInventory();
-        return FeedingLogic.IsEligibleContainer(
-            container.name,
-            inventory?.NrOfItems() ?? 0,
-            Plugin.ChestPrefix.Value
-        );
-    }
-}
