+++ Extensions/Vector3Extensions.cs
@@ -4,20 +4,30 @@ public static class Vector3Extensions
 {
     public static List<Container> GetContainersInRange(this Vector3 center, float radiusRange)
     {
-        try
+        var result = new List<Container>();
+        foreach (var piece in Piece.s_allPieces)
         {
-            Collider[] collidersInRange = Physics.OverlapSphere(
-                center,
-                Mathf.Max(radiusRange, 0),
-                LayerMask.GetMask(PluginSettings.ChestLayer)
-            );
+            if (piece == null) continue;
+            var container = piece.GetComponent<Container>();
+            if (container == null) continue;
+            var dist = Vector3.Distance(piece.transform.position, center);
+            if (dist <= radiusRange && IsValidZNetView(container.GetComponent<ZNetView>()) && IsEligibleContainer(container))
+                result.Add(container);
+        }
+        result.Sort((a, b) => Vector3.Distance(a.transform.position, center).CompareTo(Vector3.Distance(b.transform.position, center)));
+        return result;
+    }
 
-            return collidersInRange.GetContainersOrderedByDistanceFromCenter(center);
-        }
-        catch (Exception e)
-        {
-            Plugin.Log.LogWarning($"AutoFeed: Error getting containers in range: {e.Message}");
-            return [];
-        }
+    private static bool IsValidZNetView(ZNetView? zNetView) =>
+        zNetView is not null && zNetView.IsValid();
+
+    private static bool IsEligibleContainer(Container container)
+    {
+        var inventory = container.GetInventory();
+        return FeedingLogic.IsEligibleContainer(
+            container.name,
+            inventory?.NrOfItems() ?? 0,
+            Plugin.ChestPrefix.Value
+        );
     }
 }
