+++ PluginSettings.cs
@@ -2,6 +2,5 @@ namespace AutoFeed;
 
 public static class PluginSettings
 {
-    public const float FeedInterval = 0.1f;
-    public const string ChestLayer = "piece";
+    public const float FeedInterval = 5f;
 }
